+++ src/neptune/internal/init/project.py
@@ -34,14 +34,13 @@ from neptune.types.mode import Mode
 
 
 def init_project(
-    *,
     project: Optional[str] = None,
+    *,
     api_token: Optional[str] = None,
     mode: Optional[str] = None,
     flush_period: float = DEFAULT_FLUSH_PERIOD,
     proxies: Optional[dict] = None,
 ) -> Project:
-    # TODO: change * when NPT-12952 is merged
     verify_type("project", project, (str, type(None)))
     verify_type("api_token", api_token, (str, type(None)))
     verify_type("mode", mode, (str, type(None)))